Extract tests that has different behavior between NodeJS and browsers
If a node doesn't have an XML namespace but its parent has one,

- NodeJS: fast-xml-parser sets xmlns to be the same as parent's.

- Browsers: DOM API adds an empty xmlns for the node, which means its XML
namespace is the same as its parent.
